[EDIT] added gnl, removed INT_MIN|MAX macros because included in limits.h, changed ft_strcpy|strlen

+++ libft/src/ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: pauljull <pauljull@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/14 13:58:38 by pauljull          #+#    #+#             */
-/*   Updated: 2019/09/18 06:48:41 by pauljull         ###   ########.fr       */
+/*   Updated: 2019/11/27 15:41:58 by aboitier         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-int	ft_strlen(const char *s)
+size_t	ft_strlen(const char *s)
 {
-	int	index;
+	size_t	index;
 
 	if (!s)
 		return (0);
